Refactored workflow
