--- tools/enricher/README.md
@@ -1,7 +1,0 @@
-# Overview
-This enriches bounties that have been pushed to the staging branch.
-
-Enrichment includes:
-- Opening GitHub Issues (if none exist)
-- Creating GitHub Comments (if no Issue exists)
-- Creating a 418sec fork of the original repository
