--- tools/enricher/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "enricher",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@octokit/rest": "^17.10.0",
-    "mustache": "^4.0.1"
-  },
-  "engines": {
-    "node": ">14.4.0"
-  },
-  "devDependencies": {}
-}
