--- tools/indexer/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "indexer",
-  "version": "1.0.0",
-  "description": "Creates an index of all the vulnerabilities with open bounties",
-  "main": "index.js",
-  "scripts": {
-  },
-  "author": "418sec",
-  "license": "ISC",
-  "dependencies": {
-    "fdir": "^2.1.0"
-  }
-}
