+++ tools/process-bounties/assets/templates/github-issue-body.mustache
