+++ tools/process-bounties/assets/templates/github-issue-comment-body.mustache
