+++ tools/process-bounties/assets/templates/github-issue-title.mustache
