+++ tools/process-bounties/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "process-bounties",
+  "version": "2.0.0",
+  "description": "Creates an index of all the vulnerabilities with open bounties",
+  "scripts": {},
+  "author": "418sec",
+  "license": "ISC",
+  "dependencies": {
+    "@octokit/rest": "^17.11.0",
+    "fdir": "^3.4.3",
+    "mustache": "^4.0.1"
+  },
+  "engines": {
+    "node": ">14.4.0"
+  }
+}
